fix(editor): keep angle brackets inside consecutive code fences
Fence detection used ^\s* which crosses newlines, so a blank line between
fenced blocks opened a phantom fence that the next opener closed. The second
fence body was then unfenced and HTML-shaped placeholders became ORCA_RICH_MD
tokens. Match only horizontal whitespace and consume delimiter lines whole.

Fixes #13307
