+++ src/renderer/src/components/editor/raw-markdown-html.ts
@@ -1,14 +1,10 @@
-import { Node, mergeAttributes } from '@tiptap/core'
 import { isEditableDetailsHtmlBlock, matchDetailsHtmlBlock } from './details-markdown-html'
 import { formatMarkdownDocLinkBody, parseMarkdownDocLink } from './markdown-doc-links'
 import { normalizeMarkdownReferenceLinks } from './markdown-reference-link-normalization'
-import type {
-  RichMarkdownEditorCodec,
-  RichMarkdownSourceKind,
-  RichMarkdownSourceTransport
-} from './rich-markdown-source-transport'
+import type { RichMarkdownEditorCodec } from './rich-markdown-source-transport'
 import { isReservedRichMarkdownTransportBody } from './rich-markdown-source-transport'
 import { matchHtmlSuperscriptLinkSource } from './rich-markdown-html-superscript-link-source'
+import { consumeMarkdownFenceDelimiterLine } from './raw-markdown-html-fence'
 
 const INLINE_HTML_PATTERN = /^<!--[\s\S]*?-->|^<\/?[A-Za-z][\w.:-]*(?:\s[^<>]*?)?\/?>/
 
@@ -62,31 +58,21 @@ export function encodeRawMarkdownHtmlForRichEditor(
   const { transport } = codec
   let index = 0
   let isLineStart = true
-  let activeFence: '`' | '~' | null = null
-  let activeFenceLength = 0
+  const fenceState = { activeFence: null as '`' | '~' | null, activeFenceLength: 0 }
   let result = ''
-
   while (index < normalizedContent.length) {
     if (isLineStart) {
-      // Why: only line starts inspect the rest of the line, so slicing the suffix on every
-      // character (one throwaway string per char) is pure waste — compute it here. On a large
-      // doc this drops O(n) suffix allocations from the rich-editor open path (#7056).
-      const lineRest = normalizedContent.slice(index)
-      const fenceMatch = lineRest.match(/^\s*(`{3,}|~{3,})/)
-      if (fenceMatch) {
-        const fenceChar = fenceMatch[1][0] as '`' | '~'
-        const fenceLength = fenceMatch[1].length
-        if (activeFence === null) {
-          activeFence = fenceChar
-          activeFenceLength = fenceLength
-        } else if (activeFence === fenceChar && fenceLength >= activeFenceLength) {
-          activeFence = null
-          activeFenceLength = 0
-        }
+      // Why: fence open/close must not let ^\s* cross newlines (#13307 / #7056).
+      const fenceEnd = consumeMarkdownFenceDelimiterLine(normalizedContent, index, fenceState)
+      if (fenceEnd !== null) {
+        result += normalizedContent.slice(index, fenceEnd)
+        isLineStart = fenceEnd > index && normalizedContent[fenceEnd - 1] === '\n'
+        index = fenceEnd
+        continue
       }
     }
 
-    if (activeFence) {
+    if (fenceState.activeFence) {
       const nextChar = normalizedContent[index]
       result += nextChar
       isLineStart = nextChar === '\n'
@@ -214,123 +200,8 @@ export function encodeRawMarkdownHtmlForRichEditor(
   return result
 }
 
-export function createRichMarkdownLiteral(transport: RichMarkdownSourceTransport) {
-  return createRawSourceNode({
-    name: 'richMarkdownLiteral',
-    kind: 'literal',
-    inline: true,
-    transport,
-    marker: 'data-rich-markdown-literal'
-  })
-}
-
-export function createRawMarkdownHtmlInline(transport: RichMarkdownSourceTransport) {
-  return createRawSourceNode({
-    name: 'rawMarkdownHtmlInline',
-    kind: 'inline-html',
-    inline: true,
-    transport,
-    marker: 'data-raw-markdown-html-inline',
-    className: 'raw-markdown-html-inline'
-  })
-}
-
-function createRawSourceNode({
-  name,
-  kind,
-  inline,
-  transport,
-  marker,
-  className
-}: {
-  name: string
-  kind: RichMarkdownSourceKind
-  inline: boolean
-  transport: RichMarkdownSourceTransport
-  marker: string
-  className?: string
-}) {
-  return Node.create({
-    name,
-    inline,
-    group: inline ? 'inline' : 'block',
-    atom: true,
-    selectable: true,
-
-    addAttributes() {
-      return {
-        value: {
-          default: '',
-          rendered: false
-        }
-      }
-    },
-
-    // Why: converting embedded HTML tags into placeholder tokens before the
-    // markdown parser runs keeps marked's built-in paragraph tokenization intact
-    // while still letting Orca round-trip the raw markup verbatim.
-    markdownTokenName: name,
-    markdownTokenizer: {
-      name,
-      level: inline ? 'inline' : 'block',
-      start: transport.startFor(kind),
-      tokenize(src) {
-        const matched = transport.match(src, kind)
-        if (!matched) {
-          return undefined
-        }
-
-        return {
-          type: name,
-          raw: matched.raw,
-          text: matched.value,
-          block: !inline
-        }
-      }
-    },
-    parseMarkdown: (token, helpers) => {
-      if (token.type !== name) {
-        return []
-      }
-
-      return helpers.createNode(name, {
-        value: typeof token.text === 'string' ? token.text : ''
-      })
-    },
-    renderMarkdown: (node) => (typeof node.attrs?.value === 'string' ? node.attrs.value : ''),
-    renderText: ({ node }) => (typeof node.attrs.value === 'string' ? node.attrs.value : ''),
-
-    parseHTML() {
-      return [
-        {
-          tag: `${inline ? 'span' : 'div'}[${marker}]`,
-          getAttrs: (element: HTMLElement) => ({ value: element.textContent ?? '' })
-        }
-      ]
-    },
-
-    renderHTML({ HTMLAttributes, node }) {
-      const value = typeof node.attrs.value === 'string' ? node.attrs.value : ''
-      return [
-        inline ? 'span' : 'div',
-        mergeAttributes(HTMLAttributes, {
-          [marker]: '',
-          contenteditable: 'false',
-          class: className
-        }),
-        inline ? value : ['pre', value]
-      ]
-    }
-  })
-}
-
-export function createRawMarkdownHtmlBlock(transport: RichMarkdownSourceTransport) {
-  return createRawSourceNode({
-    name: 'rawMarkdownHtmlBlock',
-    kind: 'block-html',
-    inline: false,
-    transport,
-    marker: 'data-raw-markdown-html-block',
-    className: 'raw-markdown-html-block'
-  })
-}
+export {
+  createRichMarkdownLiteral,
+  createRawMarkdownHtmlInline,
+  createRawMarkdownHtmlBlock
+} from './raw-markdown-html-nodes'
